+++ apps/code-examples/src/app/code-examples/ag-grid/data-grid/basic/demo.component.ts
@@ -3,10 +3,18 @@ import { SkyAgGridModule, SkyAgGridService, SkyCellType } from '@skyux/ag-grid';
 import { SkyDataManagerService } from '@skyux/data-manager';
 
 import { AgGridModule } from 'ag-grid-angular';
-import { ColDef, GridOptions, ValueFormatterParams } from 'ag-grid-community';
+import {
+  AllCommunityModule,
+  ColDef,
+  GridOptions,
+  ModuleRegistry,
+  ValueFormatterParams,
+} from 'ag-grid-community';
 
 import { ContextMenuComponent } from './context-menu.component';
 import { AG_GRID_DEMO_DATA, AgGridDemoRow } from './data';
+
+ModuleRegistry.registerModules([AllCommunityModule]);
 
 @Component({
   selector: 'app-demo',