+++ libs/components/code-examples/src/lib/modules/ag-grid/data-grid/inline-help/example.component.ts
@@ -10,10 +10,12 @@ import { SkySearchModule } from '@skyux/lookup';
 
 import { AgGridModule } from 'ag-grid-angular';
 import {
+  AllCommunityModule,
   ColDef,
   GridApi,
   GridOptions,
   GridReadyEvent,
+  ModuleRegistry,
   ValueFormatterParams,
 } from 'ag-grid-community';
 import { of } from 'rxjs';
@@ -21,6 +23,8 @@ import { of } from 'rxjs';
 import { ContextMenuComponent } from './context-menu.component';
 import { AG_GRID_DEMO_DATA, AgGridDemoRow } from './data';
 import { InlineHelpComponent } from './inline-help.component';
+
+ModuleRegistry.registerModules([AllCommunityModule]);
 
 /**
  * @title Basic setup with inline help (without data manager)