+++ libs/components/code-examples/src/lib/modules/ag-grid/data-grid/paging/example.component.ts
@@ -12,10 +12,12 @@ import { SkyPagingModule } from '@skyux/lists';
 
 import { AgGridModule } from 'ag-grid-angular';
 import {
+  AllCommunityModule,
   ColDef,
   GridApi,
   GridOptions,
   GridReadyEvent,
+  ModuleRegistry,
   ValueFormatterParams,
 } from 'ag-grid-community';
 import { Subscription } from 'rxjs';
@@ -23,6 +25,8 @@ import { filter, map } from 'rxjs/operators';
 
 import { ContextMenuComponent } from './context-menu.component';
 import { AG_GRID_DEMO_DATA, AgGridDemoRow } from './data';
+
+ModuleRegistry.registerModules([AllCommunityModule]);
 
 /**
  * @title Basic setup with paging (without data manager)